requirements.txt
# General dependencies for your project
torch==2.0.1+cu117
torchvision==0.15.2+cu117
numpy>=1.19.2
scipy
scikit-image==0.17.2
tqdm
matplotlib
imageio
imageio-ffmpeg
Pillow
opencv-python==4.1.2.30
scikit-learn
tensorboard
tensorboardX
absl-py
accelerate
gin-config
mediapy
trimesh
pymeshlab
xatlas
plyfile
rawpy
ninja
pyhocon
dotmap
pretrainedmodels
lpips
pudb==2019.2
invisible-watermark
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.19.2
torchmetrics==0.6.0
kornia==0.6
albumentations==0.4.3
opencv-contrib-python
pyyaml
pyopengl
